feat: moving config beans to config package (#385)
* feat: moving config beans to config package

* feat: moving config beans to config in spring-boot3 example

(cherry picked from commit abf162d80cba5bfb53a5ca6333c16679b4943205)
