+++ springboot3/src/main/java/org/acme/config/L2CacheConfiguration.java
@@ -1,4 +1,4 @@
-package org.acme;
+package org.acme.config;
 
 import java.net.URI;
